restore old functionality
